refactor: make stories concise
