+++ stories/components/ButtonStory/Color/SolidButtonColors.tsx
@@ -1,11 +1,13 @@
+import {boolean, number} from '@storybook/addon-knobs';
+
+import {Button} from '../../../../main';
 import type {ButtonColorType} from '../../../../main';
-import {IconButton} from '../../../../main';
 // Caveat: Expo web needs React to be imported
 import React from 'react';
 import type {ReactElement} from 'react';
 import {StorySection} from '../../../GlobalStyles';
 
-function Colors(): ReactElement {
+function SolidButtonColors(): ReactElement {
   const colors: ButtonColorType[] = [
     'primary',
     'secondary',
@@ -19,16 +21,19 @@ function Colors(): ReactElement {
   return (
     <StorySection>
       {colors.map((color) => (
-        <IconButton
+        <Button
           key={color}
           color={color}
-          size="small"
-          icon="cross-light"
-          style={{padding: 4}}
+          disabled={boolean('disabled', false)}
+          activeOpacity={number('activeOpacity', 0.8)}
+          text={color.toUpperCase()}
+          style={{margin: 4}}
+          // eslint-disable-next-line no-console
+          onPress={() => console.log(`press ${color} solid button`)}
         />
       ))}
     </StorySection>
   );
 }
 
-export default Colors;
+export default SolidButtonColors;
