+++ stories/components/ButtonStory/Color/TextButtonColors.tsx
@@ -1,11 +1,13 @@
-import {Button} from '../../../../../main';
-import type {ButtonColorType} from '../../../../../main';
+import {boolean, number} from '@storybook/addon-knobs';
+
+import {Button} from '../../../../main';
+import type {ButtonColorType} from '../../../../main';
 // Caveat: Expo web needs React to be imported
 import React from 'react';
 import type {ReactElement} from 'react';
-import {StorySection} from '../../../../GlobalStyles';
+import {StorySection} from '../../../GlobalStyles';
 
-function OutlinedButtonColors(): ReactElement {
+function TextButtonColors(): ReactElement {
   const colors: ButtonColorType[] = [
     'primary',
     'secondary',
@@ -21,16 +23,18 @@ function OutlinedButtonColors(): ReactElement {
       {colors.map((color) => (
         <Button
           key={color}
-          type="outlined"
+          type="text"
+          disabled={boolean('disabled', false)}
+          activeOpacity={number('activeOpacity', 0.8)}
           color={color}
           text={color.toUpperCase()}
           style={{margin: 4}}
           // eslint-disable-next-line no-console
-          onPress={() => console.log(`press ${color} outlined button`)}
+          onPress={() => console.log(`press ${color} text button`)}
         />
       ))}
     </StorySection>
   );
 }
 
-export default OutlinedButtonColors;
+export default TextButtonColors;
