+++ stories/components/CheckboxStory/index.tsx
@@ -1,4 +1,6 @@
-import CheckboxBasic from './CheckboxBasic';
+import CheckboxBasic from './Basic';
+import CheckboxEndElement from './EndElement';
+import CheckboxStartElement from './StartElement';
 // Caveat: Expo web needs React to be imported
 import React from 'react';
 import {renderStory} from '../../Common';
@@ -9,4 +11,6 @@ import {withKnobs} from '@storybook/addon-knobs';
 storiesOf('Checkbox', module)
   .addDecorator(withKnobs)
   .addDecorator(withActions)
-  .add('Basic', () => renderStory(<CheckboxBasic />));
+  .add('Basic', () => renderStory(<CheckboxBasic />))
+  .add('StartElement', () => renderStory(<CheckboxStartElement />))
+  .add('EndElement', () => renderStory(<CheckboxEndElement />));
