+++ stories/components/IconButtonStory/Loading/LoadingElement.tsx
@@ -1,9 +1,9 @@
-import {IconButton, LoadingIndicator} from '../../../../../main';
+import {IconButton, LoadingIndicator} from '../../../../main';
 
 // Caveat: Expo web needs React to be imported
 import React from 'react';
 import type {ReactElement} from 'react';
-import {StorySection} from '../../../../GlobalStyles';
+import {StorySection} from '../../../GlobalStyles';
 
 function LoadingElement(): ReactElement {
   return (
