+++ stories/components/IconButtonStory/Sizes.tsx
@@ -1,0 +1,35 @@
+import {StoryContainer, StorySection, StoryTitle} from '../../GlobalStyles';
+import {boolean, number} from '@storybook/addon-knobs';
+
+import type {ButtonSizeType} from '../../../main';
+import {IconButton} from '../../../main';
+// Caveat: Expo web needs React to be imported
+import React from 'react';
+import type {ReactElement} from 'react';
+import {action} from '@storybook/addon-actions';
+
+function Sizes(): ReactElement {
+  const sizes: ButtonSizeType[] = ['small', 'medium', 'large'];
+
+  return (
+    <StoryContainer>
+      <StoryTitle>Sizes</StoryTitle>
+      <StorySection>
+        {sizes.map((size) => (
+          <IconButton
+            disabled={boolean('disabled', false)}
+            activeOpacity={number('activeOpacity', 0.8)}
+            loading={boolean('loading', false)}
+            onPress={action('onPress')}
+            key={size}
+            size={size}
+            icon="search-solid"
+            style={{padding: 4}}
+          />
+        ))}
+      </StorySection>
+    </StoryContainer>
+  );
+}
+
+export default Sizes;
