--- stories/components/IconButtonStory/components/index.ts
@@ -1,7 +1,0 @@
-export {default as Basic} from './Basic';
-export {default as Colors} from './Colors';
-export {default as Disabled} from './Disabled';
-export {default as Sizes} from './Sizes';
-export {default as Loading} from './Loading';
-export {default as LoadingElement} from './Loading/LoadingElement';
-export {default as IconElement} from './IconElement';
