+++ stories/components/IconButtonStory/index.tsx
@@ -1,4 +1,8 @@
-import IconButtonBasic from './IconButtonBasic';
+import IconButtonBasic from './Basic';
+import IconButtonColor from './Color';
+import IconButtonDisabled from './Disabled';
+import IconButtonIconElement from './IconElement';
+import IconButtonLoading from './Loading';
 // Caveat: Expo web needs React to be imported
 import React from 'react';
 import {renderStory} from '../../Common';
@@ -9,4 +13,8 @@ import {withKnobs} from '@storybook/addon-knobs';
 storiesOf('IconButton', module)
   .addDecorator(withKnobs)
   .addDecorator(withActions)
-  .add('Basic', () => renderStory(<IconButtonBasic />));
+  .add('Basic', () => renderStory(<IconButtonBasic />))
+  .add('Color', () => renderStory(<IconButtonColor />))
+  .add('Loading', () => renderStory(<IconButtonLoading />))
+  .add('Disabled', () => renderStory(<IconButtonDisabled />))
+  .add('IconElement', () => renderStory(<IconButtonIconElement />));
